수정: 키워드 저장 로직 account 이동 및 테스트 - save_like_keyword 로직 account 에서 관리 - 테스트
